Integrated: Brute Force Into CLI

+++ src/struct-compression-analyzer/src/comparison/mod.rs
@@ -110,8 +110,7 @@ impl GroupComparisonMetrics {
             lz_match_multiplier: compression_options.lz_match_multiplier,
             entropy_multiplier: compression_options.entropy_multiplier,
         }) as u64;
-        let zstd_size =
-            get_zstd_compressed_size(bytes, compression_options.zstd_compression_level) as u64;
+        let zstd_size = get_zstd_compressed_size(bytes, compression_options.zstd_compression_level);
 
         GroupComparisonMetrics {
             lz_matches,
